+++ include/model.h
@@ -29,6 +29,7 @@ struct CalcReportBundle{
   double r_probe2;
   std::vector<std::string> included_elements;
   std::string chemical_formula;
+  double molar_mass;
   // output options
   bool make_report;
   bool make_full_map;
@@ -66,8 +67,8 @@ struct Atom;
 class Space;
 class Model{
   public:
-    // radius file import
-    bool importRadiusFile(std::string&);
+    // elements file import
+    bool importElemFile(const std::string&);
     std::unordered_map<std::string, double> extractRadiusMap(const std::string&);
     // atom file import
     bool readAtomsFromFile(const std::string&, bool);
@@ -103,6 +104,7 @@ class Model{
 
     double findRadiusOfAtom(const std::string&);
     double findRadiusOfAtom(const Atom&);
+    double findWeightOfAtom(const std::string&);
 
     // controller-model communication
     CalcReportBundle generateData();
@@ -113,12 +115,11 @@ class Model{
     void setAtomListForCalculation();
     void linkAtomsToAdjacentAtoms(const double&);
     void linkToAdjacentAtoms(const double&, Atom&);
-    CalcReportBundle calcVolume();
-    bool setParameters(std::string, std::string, bool, bool, bool, bool, double, double, double, int, bool, bool, bool, std::unordered_map<std::string, double>, std::vector<std::string>);
+    bool setParameters(const std::string, const std::string, const bool, const bool, const bool, const bool, const double, const double, const double, const int, const bool, const bool, const bool, const std::unordered_map<std::string, double>, const std::vector<std::string>);
     std::vector<std::tuple<std::string, int, double>> generateAtomList();
     void setRadiusMap(std::unordered_map<std::string, double> map);
+    std::unordered_map<std::string,double> getRadiusMap();
     bool setProbeRadii(const double, const double, const bool);
-    void generateChemicalFormula();
 
     // access functions for information stored in data
     double getCalcTime(){return _data.getTime();}
@@ -143,12 +144,15 @@ class Model{
     double _cart_matrix[3][3]; // cartesian coordinates of vectors A, B, C
     std::string _space_group;
     std::unordered_map<std::string, double> _radius_map;
+    std::unordered_map<std::string, double> _elem_weight;
     std::unordered_map<std::string, int> _elem_Z;
     std::map<std::string, int> _atom_amounts;
     std::map<std::string, int> _unit_cell_atom_amounts; // stores atoms of unit cell to generate chemical formula
     std::vector<Atom> _atoms;
     Space _cell;
     double _max_atom_radius = 0;
+
+    void prepareVolumeCalc();
 };