+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This changelog keeps track of changes in a user-friendly way. It is based on [ke
 
 ### Added
 * Fully functional command line interface that allows accessing all functionalities directly form the command line.
+* Macroscopic volume and surface area values are now provided in the report.
+* Unit cell volume fractions are now provided in the report.
 
 ## [v0.1.0](https://github.com/jmaglic/MoloVol/releases/tag/v0.1.0) - 2021-06-17
 * First beta release